aes-hash: remove const AesHash
We can't easily do AES key scheduling in a const context, so supporting
a const fixed key AesHash is difficult.

To avoid this, we replace the const AesHash with a fixed_key() function,
which performs AES key scheduling on its first call.

The next commit will refactor callers to use the `fixed_key()` function.