schmivitz: update H2 implementation #4
This moifies `H2` to use an updating hasher and three challenge
functions (`H_2^j` in the spec) that take specific parameters.

The goal is to make the call sites clearer (passing typed parameters
instead of a byte array), and to make it possible to pass hash structs
(like `H3`) directly to the challenge functions.