schmivitz: rearrange first challenge gen
Now, the first challenge is generated in the Vole `create` and
`reconstruct` functions (for prover and verifier, resp.).

At this point, I'm switching to the philosophy of the FAEST spec and
assuming that each challenge is used to create the subsequent challenge
and therefore we don't need to put them in the proof itself.

- Removes the old "verifier-side vole challenge" function from the Vole
  trait (we assume it now happens in `reconstruct` on the verifier side)
- Removes the vole challenge (`chall1`) from the proof struct.