schmivitz: put final challenge gen into proof
Before, the challenge 3 generation was part of the random vole API.
However it doesn't depend on anything VOLE specific (it's used for
checking the VOLEs, but it's not created with values drawn from the
VOLE).

So, I removed it from the VOLE API and put it into the proof itself with
a fixed size.