schmivitz: remove redundant values from proof
There are a couple of things in the proof that are redundant. The FAEST
spec minimizes the proof size and updates the verification to recompute
the parts that are no longer included.
- instead of passing witness challenges (`chall2`), we just pass its
  size
- instead of passing `b~`, we recompute it (c~ = a~D + b~, where
  `a~` is in the proof, `D` comes from VOLE reconstruction, and `c~` is
  computed during verify)

This gives a few simplifications e.g. we don't need to return the
witness challenges from the prover traverser because they aren't needed
after that.