schmivitz: use F8b -> F128b superfield function
In some places, we used a manual conversion from a list of 16 F8bs to a
single element in F128b. The manual conversion used standard
bit-shifting to combine the bytes.  That conversion treats the values as
integers mod 2^128.

However, the field F128b should represent polynomials mod
x^128 + x^7 + x^2 + x + 1. The multiplicative homomorphism in that field
is not preserved with a bit-shifting conversion. Instead, we use the
`form_superfield` function to convert.

This does not replace uses from the VOLEHasah or ToField functions,
because those don't seem to be a problem yet.

- Replace uses of manual conversion function for corrections
- Remove old conversion function and corresponding test
- Remove a test that makes an incorrect assumption about the values of
  each byte after conversion.