Refactor correlated OT to use only one `delta` value
We don't actually have a use for correlated OT with a vector of
`delta` values. We _do_ have a need for correlated OT with a single
delta value (in malicious garbling), and a single `delta` does seem
like the more common use-case here. So this commit refactors things to
use a single `delta`, alongside only outputting the "zero" value in
the OT, since the one value can be derived by xor-ing in `delta`.