auth-bit: Pass bits as Vec to open
Instead of passing the AuthenticatedBits as a slice to open, we pass
them as a Vec. By doing so, we prevent the user from re-using them and
force them consumes the bits instead. Re-using these bits may result in
security leaks because their security relies on a one time pad.