crates/adversary: Add a crate for items related to adversaries.
There are currently two marker traits, SemiHonest and Malicious, defined in the
Scuttlebutt crate.
These are used (primarily in the Ocelot crate) to denote the security of OT
schemes in terms of the types of adversary they are secure against.

This new crate is part of moving individual pieces of functionality out of
Scuttlebutt, ultimately helping to keep Swanky maintainable and structured for
reuse of its core components.