crates/channel-legacy: Add crate for the legacy channel abstraction.
This is an extraction of the `scuttlebutt::channel` module, which is our
'legacy' abstraction for communication in Swanky.

The crate documentation emphasizes that this will be deprecated/removed in a
future Swanky release, and that the `swanky-channel` crate is preferred for any
new code.

Because it was convenient to do so while extracting this functionality, this
commit includes a handful of small improvements to the existing code, mostly in
the form of Clippy lints related to needless borrowing / uses of `mut`, and
simplifying some error-handling using conveniences from the standard
library (and avoiding the needless creation of explicit closure values).