fancy: Remove base_conversion
base_conversion is only used for the deprecated/insecure arithmetic
garbling protocol, and even then it's only used as an optimization.

Due to recent changes, the base_conversion code was panicking due to UB
caused by arbitrary linker alignment issues. Rather than fixing the
issue (for this deprecated code), it's easier to simply remove it.