popsicle: use UNIX domain socketpair() instead of TcpSocket
The popsicle benchmark was failing because multiple invocations of the
benchmark tried to bind() to the same address. Because there's a period
of time in between when a listener gets closed and when another one can
bind() to the same address, the benchmark was failing.

To fix this, we use a unix socketpair instead of a TcpListener.