Add types for correlation-robust hashing
Before this commit, we had a single type `AesHash` that supports both
correlation-robust and tweakable circular correlation-robust hashing.
This is potentially confusing and error prone, as generally you think
of a hash function as accomplishing _one_ of these capabilities, not
both.

Thus, this commit introduces `CorrelationRobustHash` and
`TweakableCircularCorrelationRobustHash` to separate out these concerns
into two separate types. It also improves the documentation of the
crate as a whole.