aes-hash: Make the tweak a `u128` instead of a `U8x16`
The correlation robust hashes are used in settings where the tweak is an index
(e.g., the gate number in a garbling scheme). Thus it makes more sense to
treat the tweak as an integer (i.e. a `u128`) than a block. This also
makes the API less error prone, since it's easy to accidentally swap the
input and the tweak.