`authenticated-bits`: refactor `open` to output opened bits
Before, when opening authenticated bits/shares, the actual values that were
authenticated were never returned. However, there are settings where the user
would want such information (generally, you'd like to know what was authenticated
at some point!).

This commit adds an argument to the two `open` methods for authenticated
bits and shares that contains an `output: &mut Vec<F2>` in which the output
bits / combined shares are placed.

In addition, it removes the returned `bool` signifying whether the bits
were opened correctly or not, instead moving that into the error condition.
This is less error-prone, as it is no longer up to the caller to validate
themselves that opening worked as intended.