`authenticated-bits`: Add (preliminary/insecure) support for AND triples
This commit adds AND triples to the authenticated bits arsenal. AND triples
comprise three authenticated shares that form an AND relationship (i.e., for
shares `x`, `y`, and `z` it holds that `x y = z`).

AND triples are constructed by first generating _leaky_ AND triples, and then
bucketing these to avoid the leakage. Leaky AND triples are the same as normal
AND triples except an adversary could "guess" a bit and not get caught if
their guess is correct.

There are still two places with gaps in the implementation:
- The use of the `F_eq` functionality isn't included yet. There is a separate
  branch with that change, once merged in it'll be added here.
- We need to permute the leaky AND triples before bucketing. That step hasn't
  been completed yet but there's no blocker.