`authenticated-bits`: Remove spurious test
This commit removes the `wrong_generators_fail` test. This turned out to not be
a meaningful test that was causing spurious test failures.

In more detail, the test was verifying that the wrong generators cause
`AuthShare` validation to fail. A wrong generator really just means that
the wrong `delta` value is being used. _However_, when validating a zero bit,
`delta` doesn't come into play, so even _if_ a wrong `delta` is used,
verification will succeed against a zero bit! So, this test really wasn't
testing anything meaningful, and we already have a test in `authbits` that
tests against an invalid `delta`. Hence, we remove this test.