serialization: Support CanonicalSerialize for arrays
Note that `serde` only implements serialization for arrays up to 32
elements long. Because `CanonicalSerialize` requires `serde::Serialize`,
we inherit this requirement. (We could lift this requirement if
desired; I think it's only used in one place.)

The types are also _unfun_ (tho users shouldn't notice this), because
Rust const support is still nascent. `CanonicalSerialize` requires that
we have a serialization size known at compile-time. But Rust won't let
us use a const in a trait as the length of an array (hence the use of
GenericArray). GenericArray operates on typenums, so we need to use the
multiplication implemented the typesystem.

All of this to just say that `sizeof([T; N]) = sizeof(T) * N`

Like I said, users shouldn't notice this.