etc/ci/target_dir_cache: Add cache design as module docs.
This is to preserve the reasoning for the cache's architecture in the code
implementing that architecture, rather than keeping it entirely separate on the
Swanky internal wiki (this is useful for future maintainers and downstream users
that may wish to replicate the Swanky CI setup).

This module comment uses reST syntax rather than Markdown, as this is more
common for Python documentation (since it can be rendered with e.g. Sphinx).