.github/workflows/normal.yml: Set up GitHub workflow for basic CI.
Due to our GitHub repository's configuration as a push mirror for only protected
branches of our internal repository, we only need to run CI tasks in a few
scenarios:

- Manually, so it's easy to test on the public runners
- On pushes to the `dev` branch (which correspond to feature merges in our
- internal repository)
- On pushes to tags (for our tagged releases in particular)

We run our test job on both x86 and ARM64 runners, and otherwise the structure
is (more or less) the same as the GitLab version of normal tests in
`.gitlab-ci.yml`.

Notably, however, we don't get any caching benefit - this requires access to a
persistent filesystem.
It may be worth considering GitHub's cache action, but this has very limited
storage, and would have limited benefit given what triggers CI on GitHub.