Delete the .github directory and workflows.
We discovered quickly that the public GitHub runners are not sufficiently large
to run ./swanky ci quick.

Rather than trying to get this working as GitHub actions, we are instead working
on simply publishing our internal pipeline results to GitHub using our internal
repository host's GitHub integration.
With this, it is no longer necessary to configure any actions to run.