`authenticated-bits`: Remove `bitxor_assign` impl from `AuthBits`
The `bitxor_assign` implementation for `AuthBits` was _incorrect_. This is
because there was a copy happening under-the-hood, making the mutable change
happen to the _copied_ value, not the original value. Hence we remove this
implementation for now. We'd need some way to extract a mutable reference from
a `PartyEither` to allow `bitxor_assign` to work properly.