`authenticated-bits`: Improve performance of combining leaky triples
This commit improves the performance of combining leaky triples by 50%.
It does so by batching the opening of the `d` values when bucketing into
a single `AuthSharesGenerator::open` call (versus an `open` call per bucket).
Everything else in the `combine` method is local, so this reduces the rounds
of communication from `nbuckets` to `1`, which is where the 50% performance
improvement comes from.