`authenticated-bits`: Remove generic OT trait from authenticated-* definitions
The authenticated bit/share/etc generators were generic in their use of the
underlying OT. However, this type of generality is not useful, as certain
optimizations show that generic OT is not needed, but rather a more specialized
functionality can be used. This commit moves us one step towards implementing
this optimizations by reducting the generic OT and simply hardcoded the KOS
construction. A future commit will replace this with something more efficient.