`authenticated-bits`: Hopefully clearer presentation of leaky triple generation
This commit adds extensive documentation, and also some optimizations made clear
with the new documentation, to the leaky and triple generation method. The goal
is that this code should be mostly understandable to someone reading the comments
(although I did not explain the half-gates details due to not wanting to overload
the documentation _too_ much).