`authenticated-bits`: Pass interator to `AuthBitGenerator::generate`
Previously, `AuthBitGenerator::generator` took a slice of `F2` elements.
However, this forces the caller to have the `F2` elements in a vector, and
can't, for instance, use a more compact representation for these elements.
Instead, we replace the `&[F2]` with an `Iterator<Item = F2>` to allow for
more flexibility in the API.