Disentangle SIEVE-IR parsing from ZK execution by using Circuit
Before this commit the parsing of SIEVE-IR was done together with the circuit evaluation.
This commit separate the two tasks which allows for better performance and facilitate benchmarks.
This commit also replaces the HashMap used for circuit evaluation with the new MemoryCircuit from
the circuit module. This replacement provides a 2x improvement by using a Vec instead of a HashMap.
One downside is a lot of the tests are now obsolete because the MemoryCircuit does not provide
the catch the same errors like overwrite a wire in memory.