party2::ty_eq: Clarify what `Generic` wrappers give us.
Previously, the comment mentions that we can use them for "type
conversions".
This is somewhat unclear given that Rust already has a notion of
casting (both with `as` and via the `From`/`Into`/`TryFrom`/`TryInto`
traits), and doesn't really 'fit' the code that follows in an obvious
way.

The new comment clarifies that the following functions, `convert_set`
and `convert_map`, provide a way to turn evidence that `T0 == T1` into
evidence of the corresponding equality on the complex generic types.

The notion of type conversion previously alluded to in the
documentation is later explicated in the documentation of the `cast`
method on `EqualityProposition`.