party2::ty_eq: Add extended module documentation.
This module represents a fairly complicated set of abstractions that
reflect some of the type equality reasoning that exists in the Rust
compiler / type-checker into our own code, so that we can take
advantage of some clever tricks that allow the use of a `PartySystem`
to be as inexpensive (and ergonomic) as possible.

The added documentation walks through some of the underlying theory,
building up context for the 'core' module types/traits:
`EqualityProposition`, `TrueEqualityProposition`,
`NotNecessarilyTrueEqualityProposition`, and `Witness`.

Discussion of `Generic` and the handling of disjunctive propositions
was left for those types' respective documentation blocks, as they add
some additional complexity on top of the already-complex idea of
reflecting type-level ideas at the value level (and vice-versa)
without a real notion of dependent types.

Given that this module is not intended to be used _super_ extensively,
this piece of documentation is more informational, rather than
tutorial-like.