party2::ty_eq: Clarify what is meant by 'sealing' a trait.
And, the effects on users of the module that this has.

This clarification is accomplished in two ways:

- Linking to Rust's internal API guidelines, which use this pattern to
  prevent changes to a trait from being _breaking_ changes, which
  would force a major version increase in SEMVER.
- Removing the use of the word 'downstream', instead referring to
  'code using this module'.

Another important motivation -- preventing the introduction of
inconsistency/unsoundness to the logic encapsulated by the trait -- is
given proper space, rather than existing as a restatement of what
'misuse of the trait' meant in the language used previously.