party2::ty_eq: Reword description of `TrueEqualityProposition`.
Previously, the sentence ended rather abruptly and unclearly.

This commit clarifies that this type uses its single value (being a
structure type that acts like `()`) as the witness it is able to
summon.
This language matches better with the description of
`NotNecessarilyTrueEqualityProposition`, which discusses that no
witness _can_ be summoned since that type has no constructors.