auth-gc: Add a fancy object which counts gates
This commit adds a a fancy object called Analyzer which keeps track of
the number of each type of gate (e.g. number of ANDs). The code itself
still needs some documentation and better tests. The Analyzer object
does not compute any part of the circuit and simply traverses it for the
sake of profiling it.

We will eventually use the Analyzer in order to figure out how to
pre-process a circuit for authenticated garbling.