party2: Add top-level module documentation.
This is intended to introduce the `party2` crate in the same way
`party` is introduced (and, to an extent, describe the new features
that are possible with `party_map`).

It follows essentially the same structure as the original party
crate's top-level documentation, as this succinctly demonstrates the
value and use of the abstraction.

It is also the case that party2 is _nearly_ a drop-in replacement, but
slightly more robust / capable when it comes to complex
implementations involving party code (even those involving complicated
data structures).